move files inside `emission_factors`
