+++ bedrock/publish/emission_factors/placeholders.py
